add new students

+++ database/migrations/2024_10_02_1320234_create_teachers_table.php
@@ -11,16 +11,18 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('teachers', function (Blueprint $table) {
-            $table->id();
-            $table->string('name');
-            $table->string('email')->unique();
-            $table->string('phone');
-            $table->string('address')->nullable();
-            $table->string('specialization')->nullable();
-            $table->foreignId('department_id')->constrained()->onDelete('cascade'); // Add department relationship
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('teachers')) {
+            Schema::create('teachers', function (Blueprint $table) {
+                $table->id();
+                $table->string('name');
+                $table->string('email');
+                $table->string('phone');
+                $table->string('address')->nullable();
+                $table->string('specialization')->nullable();
+                $table->foreignId('department_id')->constrained()->onDelete('cascade');
+                $table->timestamps();
+            });
+        }
     }
 
     /**
